pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hippius-s3"
version = "0.0.2"
description = "S3 Gateway for Hippius' IPFS storage"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "eatmyveggies", email = "veggies@hippius.com"},
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.23.0",
    "asyncpg>=0.28.0",
    "pydantic>=2.3.0",
    "python-dotenv>=1.0.0",
    "hippius",
    "httpx>=0.27.0",
    "aiofiles>=23.1.0",
    "lxml>=4.9.3",
    "botocore>=1.38.25",
    "boto3>=1.40.61",
    "hippius[key_storage]>=0.2.68",
    "cachetools",
    "redis",
    "pyinstrument>=4.0.0",
    "PyNaCl>=1.6.0",
    "substrate-interface>=1.7.11",
    "watchfiles>=1.1.0",
    "pandas>=2.3.3",
    "pyarrow>=18.0.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-instrumentation>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-redis>=0.58b0",
    "opentelemetry-instrumentation-asyncpg>=0.58b0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-exporter-prometheus>=0.58b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "prometheus-client>=0.23.1",
    "loki-logger-handler>=1.1.2",
]

[project.optional-dependencies]
proxy = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.23.0",
    "httpx>=0.27.0",
    "asyncpg>=0.28.0",
    "redis",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.12.0",
    "mypy==1.10.0",
    "pre-commit>=3.5.0",
    "minio>=7.1.15",
    "boto3>=1.34.0",
    "types-boto3",
    "types-botocore",
    "types-requests",
    "types-redis",
    "requests>=2.31.0",
    "psycopg[binary]>=3.1",
    "fakeredis>=2.23.0",
    "respx>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["hippius_s3"]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = ["examples/"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "F",        # pyflakes
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "C90",      # maccbe
    "I",        # isort
    "ASYNC",    # flake8-async
    "B",        # flake8-bugbear
    "C4",       # flake8-comprehensions
    "DTZ",      # flake8-datetimetz
    "T10",      # flake8-debugger
    "FA",       # flake8-future-annotations
    "T20",      # flake8-print
    "Q",        # flake8-quotes
    "SIM",      # flake8-simplify
    "PTH",      # flake8-use-pathlib
    "ERA",      # eradicate
    "RET",
    "PERF",     # perflint
    "TID",      # flake8-tidy-imports
]
ignore = [
    "C901", # Too complex functions
    "E722", # Bare except clauses
    "E501", # Line too long (let formatter handle this)
    "B008",
]


[tool.ruff.lint.isort]
force-single-line   = true
known-first-party = ["hippius_s3"]
known-third-party = ["hippius_sdk", "hippius"]
lines-after-imports = 2

[tool.ruff.lint.mccabe]
max-complexity = 13 # TODO: get this down to 8

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"examples/*.py" = ["T201", "PTH"]
"tests/**/*.py" = [
    "T201",      # print statements
    "ERA001",    # commented out code
    "C901",      # complex functions
    "PLR0913",   # too many arguments
    "PLR2004",   # magic values
    "PLR0915",   # too many statements
    "B011",      # assert false
    "SIM",       # simplify checks
    "PERF",      # performance checks
    "RET",       # return statement checks
]
"hippius_s3/scripts/*.py" = ["T201", "PTH"]
"scripts/*.py" = ["T201", "PTH", "SIM105"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
plugins = [
    "pydantic.mypy"
]
# Ignore errors in specific modules
[[tool.mypy.overrides]]
module = "hippius_s3.api.s3.multipart"
disallow_untyped_defs = false
disable_error_code = ["attr-defined", "no-untyped-def", "misc", "return-value"]


[[tool.mypy.overrides]]
module = "hippius_s3.queue"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def", "misc", "arg-type"]

[[tool.mypy.overrides]]
module = "hippius_s3.utils"
disallow_untyped_defs = false
disable_error_code = ["var-annotated", "no-untyped-def", "no-any-return"]

[[tool.mypy.overrides]]
module = "hippius_s3.ipfs_service"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest_asyncio"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "loki_logger_handler.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "nacl.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "hippius_s3.api.middlewares.profiler"
disallow_untyped_defs = false
disable_error_code = ["attr-defined", "no-untyped-def"]

[[tool.mypy.overrides]]
module = "hippius_s3.dependencies"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "hippius_s3.api.user"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "hippius_s3.api.middlewares.credit_check"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def", "no-any-return"]

[[tool.mypy.overrides]]
module = "hippius_s3.workers.uploader"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def", "abstract"]

[[tool.mypy.overrides]]
module = "hippius_s3.main"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "aiofiles.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lxml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "hippius_sdk.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cachetools.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mnemonic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fakeredis.aioredis"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "substrateinterface.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=hippius_s3"
markers = [
    "unit: Unit tests (isolated functions, no external dependencies)",
    "integration: Integration tests (worker logic with real DB/Redis, mocked external services)",
    "e2e: End-to-end tests (full API testing from client perspective)",
]

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prometheus_client.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def", "attr-defined", "arg-type", "return-value", "misc"]
